project_api/requirements.txt
fastapi
slowapi
uvicorn[standard]
pydantic
pydantic[email]
alembic
sqlalchemy
psycopg2-binary
asyncpg
asyncio
argon2-cffi
jwt
python-jose[cryptography]
